tools_hf/HF-GUI/requirements.txt
matplotlib>=3.5.0
requests>=2.28.0
pandas>=1.4.0
python-dotenv>=0.19.0
tkcalendar>=1.6.0
pyperclip>=1.8.0
PyJWT>=2.4.0